Streamline SQLite BLOB interface
The functions soci::blob::write and soci::blob::append were returning
the new total buffer size when using SQLite. However, on other backends,
these functions return the amount of bytes written into the BLOB.

If the implementation of Postgresql and Firebird can be believed to be
correct, then the implementation of SQLite's BLOB actually broke the
interface of soci::blob.

This discrepancy is removed with this commit. With it, SQLite will now
also return the bytes written instead of new total size.