Request SQLite to copy BLOB data
Before we were using SQLITE_STATIC indicating that we'd make sure that
the passed pointer remains valid until the processing is completely
done. However, since at this point in the code, we don't own the
pointer, we can't make any guarantees about its lifetime. Thus, we now
use SQLITE_TRANSIENT, which will cause SQLite to copy the buffer inside
the sqlite3_bind_blob call.

Note: This makes the handling of BLOBs consistent with what is already
done for strings.