Unify handling of non-supported BLOBs
Before, most backends that don't support BLOBs had dummy implementations
for all blob-related functions that simply did nothing (and returned
zero for the amount of read/written data). The exception was the MySQL
backend which would actually throw exceptions when any of the blob
functions were used.

The latter approach seems to be the more sensible one as that clearly
communicates to the user that BLOBs are not supported for this backend
whereas the former approach might send the user on a large debugging
tour, before checking the SOCI source code only to find that BLOBs
aren't actually supported for this specific backend.