Oracle: Streamline BLOB interface
This commit adapts the oracle BLOB interface such that it now is en par
with the implementations for the other backends. Notably this means that
it is now possible to perform read and write operations on a
default-constructed soci::blob object and then insert this object into
the database at a later point (instead of first having to select the
BLOB out of the database, before being able to access it).